pyproject.toml
[project]
name = "efr-maze-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "ipdb>=0.13.13",
    "labyrinth-py>=1.0.4",
]
